Revert "Merged Develop into Master (#33)"
This reverts commit bd53446c8e84c986d66a5956bebc24e88fbc5125.

+++ Drivers/ScanLab/Implementation/libmcdriver_scanlab_rtcrecordinginstance.cpp
@@ -524,8 +524,8 @@ int32_t * CRTCRecordingChannel::reserveDataBuffer(uint32_t nCount, uint32_t& nEn
 
 	if (nCount == 0)
 		throw ELibMCDriver_ScanLabInterfaceException(LIBMCDRIVER_SCANLAB_ERROR_DATARECORDINGUNDERFLOW, "data recording interval underflow");
-	if ((size_t)nCount > m_nChunkSize) {
-		nCount = (uint32_t)m_nChunkSize;		
+	if (nCount > m_nChunkSize) {
+		nCount = m_nChunkSize;		
 	}
 
 	if (m_pCurrentChunk.get() != nullptr) {
